third_party/WebKit/LayoutTests/media/viewport-in-standalone-media-document-expected.txt

This tests that a standalone media document has viewport settings.


EXPECTED (metaElement.name == 'viewport') OK
EXPECTED (metaElement.content == 'width=device-width') OK
END OF TEST
